Refactor curriculum data structure and improve backend API consistency
- Move curriculum unit definitions into separate per-level modules under app/data/en/
- Create shared types module app/data/en/_types.py for CurriculumUnit and related types
- Update app/data/en/curriculum.py to re-export types and assemble full curriculum from per-level modules
- Add health router with /health endpoint for system checks and integrate into main app
- Replace direct Redis and DB session usage with context-managed utilities in multiple routers and services
- Standardize HTTPException status codes across routers to use fastapi.status constants
- Refactor conversation websocket quota checks to use unified check_all_quotas function
- Improve ConversationPipeline TTS handling with helper methods and detailed stats tracking
- Extract email sending logic into a shared _dispatch function in email_service.py
- Add pagination utility and apply it in admin and feedback routers for consistent pagination handling
- Add async context managers for DB sessions and Redis clients in utils/db.py and utils/redis.py
- Remove redundant local Redis dependencies and replace with shared redis_client context manager
- Update imports and dependencies to use new utilities and improve code clarity
- Adjust tests environment variable SECRET_KEY value for CI consistency
- Fix .gitignore entry for /data/ directory to be correctly ignored
