+++ backend/app/data/en/curriculum_a1.py
@@ -1,0 +1,150 @@
+"""A1 curriculum units."""
+
+from __future__ import annotations
+
+from app.data.en._types import CurriculumUnit
+
+A1_UNITS: list[CurriculumUnit] = [
+    CurriculumUnit(
+        id="a1-unit-1",
+        level="A1",
+        unit_number=1,
+        title="Identity & Greetings",
+        grammar_points=["to-be", "subject-pronouns", "questions-yes-no"],
+        vocabulary_set_ids=["identity_a1", "greetings_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Introduce yourself (name, age, nationality)",
+            "Ask and answer simple personal questions",
+            "Understand basic greeting phrases",
+        ],
+        default_weeks=2,
+        prerequisite_unit=None,
+    ),
+    CurriculumUnit(
+        id="a1-unit-2",
+        level="A1",
+        unit_number=2,
+        title="My World",
+        grammar_points=["articles", "possessive-adjectives", "present-simple"],
+        vocabulary_set_ids=["family_a1", "colours_a1", "adjectives_basic_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Describe family members and possessions",
+            "Use a/an/the correctly in simple sentences",
+            "Name common colours and basic adjectives",
+        ],
+        default_weeks=2,
+        prerequisite_unit="a1-unit-1",
+    ),
+    CurriculumUnit(
+        id="a1-unit-3",
+        level="A1",
+        unit_number=3,
+        title="Daily Life",
+        grammar_points=["present-simple", "questions-yes-no"],
+        vocabulary_set_ids=["daily_routines_a1", "time_expressions_a1", "verbs_basic_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Talk about daily routines using present simple",
+            "Tell the time and use time expressions",
+            "Ask and answer yes/no questions about daily life",
+        ],
+        default_weeks=2,
+        prerequisite_unit="a1-unit-2",
+    ),
+    CurriculumUnit(
+        id="a1-unit-4",
+        level="A1",
+        unit_number=4,
+        title="Places & Location",
+        grammar_points=["there-is-are", "prepositions-place"],
+        vocabulary_set_ids=["home_a1", "city_places_a1", "prepositions_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Describe where things are using prepositions",
+            "Use there is/are to describe places",
+            "Name rooms in a house and places in a city",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a1-unit-3",
+    ),
+    CurriculumUnit(
+        id="a1-unit-5",
+        level="A1",
+        unit_number=5,
+        title="Actions Right Now",
+        grammar_points=["present-continuous"],
+        vocabulary_set_ids=["action_verbs_a1", "clothes_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Describe what is happening right now",
+            "Distinguish present simple from present continuous",
+            "Name clothes and common actions",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a1-unit-4",
+    ),
+    CurriculumUnit(
+        id="a1-unit-6",
+        level="A1",
+        unit_number=6,
+        title="Yesterday",
+        grammar_points=["past-simple"],
+        vocabulary_set_ids=[
+            "past_time_expressions_a1",
+            "regular_verbs_past_a1",
+            "irregular_verbs_basic_a1",
+        ],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Talk about completed actions in the past",
+            "Use regular and common irregular past forms",
+            "Use past time expressions (yesterday, last week, ago)",
+        ],
+        default_weeks=2,
+        prerequisite_unit="a1-unit-5",
+    ),
+    CurriculumUnit(
+        id="a1-unit-7",
+        level="A1",
+        unit_number=7,
+        title="Abilities & Wishes",
+        grammar_points=["can-cant"],
+        vocabulary_set_ids=["abilities_a1", "free_time_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Express ability and inability with can/can't",
+            "Ask for and give permission",
+            "Talk about hobbies and free-time activities",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a1-unit-6",
+    ),
+    CurriculumUnit(
+        id="a1-unit-8",
+        level="A1",
+        unit_number=8,
+        title="A1 Consolidation",
+        grammar_points=[
+            "to-be",
+            "subject-pronouns",
+            "articles",
+            "possessive-adjectives",
+            "present-simple",
+            "present-continuous",
+            "past-simple",
+            "can-cant",
+            "questions-yes-no",
+        ],
+        vocabulary_set_ids=["food_drinks_a1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Consolidate all A1 grammar with mixed practice",
+            "Read and understand simple everyday texts",
+            "Write a short paragraph about yourself",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a1-unit-7",
+    ),
+]
