+++ backend/app/data/en/curriculum_a2.py
@@ -1,0 +1,149 @@
+"""A2 curriculum units."""
+
+from __future__ import annotations
+
+from app.data.en._types import CurriculumUnit
+
+A2_UNITS: list[CurriculumUnit] = [
+    CurriculumUnit(
+        id="a2-unit-1",
+        level="A2",
+        unit_number=1,
+        title="The Recent Past",
+        grammar_points=["past-simple"],
+        vocabulary_set_ids=["irregular_verbs_a2", "past_time_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Uses 40+ common irregular past forms correctly",
+            "Narrates a sequence of past events in a short paragraph",
+            "Asks past questions using Did, Where did, When did",
+        ],
+        default_weeks=2,
+        prerequisite_unit=None,
+    ),
+    CurriculumUnit(
+        id="a2-unit-2",
+        level="A2",
+        unit_number=2,
+        title="Plans & Future",
+        grammar_points=["going-to-future", "will-future"],
+        vocabulary_set_ids=["future_plans_a2", "weather_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Distinguishes going to (planned intention) from will (spontaneous decision)",
+            "Makes predictions about the weather using will",
+            "Describes personal plans for the week using going to",
+        ],
+        default_weeks=2,
+        prerequisite_unit="a2-unit-1",
+    ),
+    CurriculumUnit(
+        id="a2-unit-3",
+        level="A2",
+        unit_number=3,
+        title="Comparisons",
+        grammar_points=["comparatives-superlatives"],
+        vocabulary_set_ids=["adjectives_comparison_a2", "places_comparison_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Compare two or more people, places, or things",
+            "Use -er/-est and more/most correctly",
+            "Describe places using superlatives",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a2-unit-2",
+    ),
+    CurriculumUnit(
+        id="a2-unit-4",
+        level="A2",
+        unit_number=4,
+        title="Ability & Permission",
+        grammar_points=["can-cant", "could-past-ability"],
+        vocabulary_set_ids=["abilities_sports_a2", "school_work_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Express past ability using could",
+            "Ask for and grant permission politely",
+            "Use may/can/could in requests",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a2-unit-3",
+    ),
+    CurriculumUnit(
+        id="a2-unit-5",
+        level="A2",
+        unit_number=5,
+        title="Quantity & Shopping",
+        grammar_points=["countable-uncountable", "some-any-much-many"],
+        vocabulary_set_ids=["food_shopping_a2", "money_prices_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Use some/any/much/many with countable and uncountable nouns",
+            "Follow a shopping conversation",
+            "Write a simple shopping list or order",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a2-unit-4",
+    ),
+    CurriculumUnit(
+        id="a2-unit-6",
+        level="A2",
+        unit_number=6,
+        title="Health & Body",
+        grammar_points=["modal-verbs", "imperatives"],
+        vocabulary_set_ids=["body_health_a2", "symptoms_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Uses should / shouldn't for advice",
+            "Uses must / mustn't for obligation and prohibition",
+            "Describes symptoms and gives health advice in a dialogue",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a2-unit-5",
+    ),
+    CurriculumUnit(
+        id="a2-unit-7",
+        level="A2",
+        unit_number=7,
+        title="Travel & Transport",
+        grammar_points=["prepositions-time", "adverbs-manner"],
+        vocabulary_set_ids=["transport_a2", "travel_a2", "directions_a2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Uses at, on, in for time expressions (at 3pm, on Monday, in June)",
+            "Uses adverbs of manner correctly (quickly, carefully, loudly)",
+            "Gives and follows directions using transport vocabulary",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a2-unit-6",
+    ),
+    CurriculumUnit(
+        id="a2-unit-8",
+        level="A2",
+        unit_number=8,
+        title="A2 Consolidation",
+        grammar_points=[
+            "past-simple",
+            "going-to-future",
+            "will-future",
+            "comparatives-superlatives",
+            "countable-uncountable",
+            "modal-verbs",
+        ],
+        vocabulary_set_ids=[
+            "irregular_verbs_a2",
+            "adjectives_a2",
+            "food_shopping_a2",
+            "transport_a2",
+            "body_health_a2",
+        ],
+        lesson_types=["reading", "writing", "review"],
+        competency_checklist=[
+            "Reads and understands a text (150–200 words) on a familiar topic",
+            "Writes a short email or message (80–100 words) using past and future tenses",
+            "Demonstrates all A2 grammar in a mixed-tense writing task",
+        ],
+        default_weeks=1,
+        prerequisite_unit="a2-unit-7",
+    ),
+]
