+++ backend/app/data/en/curriculum_b1.py
@@ -1,0 +1,144 @@
+"""B1 curriculum units."""
+
+from __future__ import annotations
+
+from app.data.en._types import CurriculumUnit
+
+B1_UNITS: list[CurriculumUnit] = [
+    CurriculumUnit(
+        id="b1-unit-1",
+        level="B1",
+        unit_number=1,
+        title="Perfect Aspects",
+        grammar_points=["present-perfect", "present-perfect-continuous", "past-perfect"],
+        vocabulary_set_ids=["time_expressions_b1", "life_changes_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Use present perfect simple vs continuous correctly",
+            "Describe events leading up to a past moment with past perfect",
+            "Choose the correct perfect aspect in context",
+        ],
+        default_weeks=2,
+        prerequisite_unit=None,
+    ),
+    CurriculumUnit(
+        id="b1-unit-2",
+        level="B1",
+        unit_number=2,
+        title="Conditionals 1 & 2",
+        grammar_points=["first-conditional", "second-conditional", "zero-conditional"],
+        vocabulary_set_ids=["conditionals_vocab_b1", "hypothetical_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Form and use zero, first, and second conditionals correctly",
+            "Distinguish real from hypothetical conditions",
+            "Write sentences speculating about the future or imagined situations",
+        ],
+        default_weeks=2,
+        prerequisite_unit="b1-unit-1",
+    ),
+    CurriculumUnit(
+        id="b1-unit-3",
+        level="B1",
+        unit_number=3,
+        title="Passive Voice",
+        grammar_points=["passive-voice-simple", "passive-voice-perfect"],
+        vocabulary_set_ids=["processes_b1", "news_events_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Form passive sentences in present and past tenses",
+            "Choose when to use passive over active voice",
+            "Describe processes and news events using the passive",
+        ],
+        default_weeks=2,
+        prerequisite_unit="b1-unit-2",
+    ),
+    CurriculumUnit(
+        id="b1-unit-4",
+        level="B1",
+        unit_number=4,
+        title="Relative Clauses & Connectors",
+        grammar_points=["relative-clauses", "discourse-connectors-b1"],
+        vocabulary_set_ids=["connectors_b1", "descriptions_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Use defining and non-defining relative clauses",
+            "Connect ideas with although/however/therefore/as a result",
+            "Write a structured paragraph with clear links between ideas",
+        ],
+        default_weeks=2,
+        prerequisite_unit="b1-unit-3",
+    ),
+    CurriculumUnit(
+        id="b1-unit-5",
+        level="B1",
+        unit_number=5,
+        title="Modals & Advice",
+        grammar_points=["modal-verbs", "should-ought-to", "must-have-to"],
+        vocabulary_set_ids=["advice_b1", "obligation_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Use must/have to/should/ought to correctly",
+            "Give and respond to advice",
+            "Express obligation, prohibition, and recommendation",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b1-unit-4",
+    ),
+    CurriculumUnit(
+        id="b1-unit-6",
+        level="B1",
+        unit_number=6,
+        title="Reported Speech",
+        grammar_points=["reported-speech"],
+        vocabulary_set_ids=["reporting_verbs_b1", "news_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Report what someone said with correct tense backshift",
+            "Use reporting verbs (say, tell, ask, explain)",
+            "Summarise a short news item or conversation",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b1-unit-5",
+    ),
+    CurriculumUnit(
+        id="b1-unit-7",
+        level="B1",
+        unit_number=7,
+        title="Wishes & Regrets",
+        grammar_points=["wish-if-only", "third-conditional"],
+        vocabulary_set_ids=["feelings_regret_b1", "hypothetical_past_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Express wishes about present and past using wish/if only",
+            "Use third conditional to talk about regrets",
+            "Write about a decision you would change",
+        ],
+        default_weeks=2,
+        prerequisite_unit="b1-unit-6",
+    ),
+    CurriculumUnit(
+        id="b1-unit-8",
+        level="B1",
+        unit_number=8,
+        title="B1 Consolidation",
+        grammar_points=[
+            "present-perfect",
+            "first-conditional",
+            "second-conditional",
+            "passive-voice-simple",
+            "relative-clauses",
+            "modal-verbs",
+            "reported-speech",
+        ],
+        vocabulary_set_ids=["consolidation_b1"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Consolidate all B1 grammar in mixed tasks",
+            "Read and understand intermediate informational texts",
+            "Write a coherent piece of 100–150 words",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b1-unit-7",
+    ),
+]
