+++ backend/app/data/en/curriculum_b2.py
@@ -1,0 +1,158 @@
+"""B2 curriculum units."""
+
+from __future__ import annotations
+
+from app.data.en._types import CurriculumUnit
+
+B2_UNITS: list[CurriculumUnit] = [
+    CurriculumUnit(
+        id="b2-unit-1",
+        level="B2",
+        unit_number=1,
+        title="Past Narratives & Sequencing",
+        grammar_points=["past-perfect"],
+        vocabulary_set_ids=["academic_vocabulary_b2", "narrative_time_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Uses the past perfect to sequence two past events correctly",
+            "Does not use the past perfect when the sequence is already clear from context",
+            "Uses time connectors (prior to, subsequently, meanwhile) to organise a narrative",
+            "Writes a coherent 150-word past narrative with correct verb forms",
+            "Distinguishes past simple (simple past action) from past perfect (earlier action)",
+        ],
+        default_weeks=1,
+        prerequisite_unit=None,
+    ),
+    CurriculumUnit(
+        id="b2-unit-2",
+        level="B2",
+        unit_number=2,
+        title="Wishes & Regrets",
+        grammar_points=["wishes-regrets"],
+        vocabulary_set_ids=["emotions_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            'Uses "I wish + past simple" for present wishes correctly',
+            'Uses "I wish + past perfect" for past regrets correctly',
+            'Uses "I wish + would" to express a complaint about behaviour',
+            'Uses "if only" for emphasis without changing the underlying structure',
+            "Expresses nuanced emotions (remorse, yearning, nostalgia) in writing",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b2-unit-1",
+    ),
+    CurriculumUnit(
+        id="b2-unit-3",
+        level="B2",
+        unit_number=3,
+        title="Conditionals 2 & 3",
+        grammar_points=["second-conditional", "third-conditional"],
+        vocabulary_set_ids=["workplace_b2", "hypothetical_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            'Forms the second conditional without "would" in the if-clause',
+            "Forms the third conditional with past perfect + would have correctly",
+            'Uses "could/might" in the main clause as alternatives to "would"',
+            "Speculates using hypothetical language (scenario, likelihood, presumably)",
+            "Does not confuse second and third conditional time frames",
+        ],
+        default_weeks=2,
+        prerequisite_unit="b2-unit-2",
+    ),
+    CurriculumUnit(
+        id="b2-unit-4",
+        level="B2",
+        unit_number=4,
+        title="Advanced Passive & Causative",
+        grammar_points=["advanced-passive"],
+        vocabulary_set_ids=["industries_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            'Forms causative "have/get + object + past participle" correctly',
+            "Uses modal passive (must be, should be, can be + past participle)",
+            "Uses passive infinitives (to be + past participle) after verbs like want, expect",
+            "Describes industrial or manufacturing processes using the passive",
+            'Selects "have" (formal) vs "get" (informal) appropriately in causative structures',
+        ],
+        default_weeks=1,
+        prerequisite_unit="b2-unit-3",
+    ),
+    CurriculumUnit(
+        id="b2-unit-5",
+        level="B2",
+        unit_number=5,
+        title="Gerunds & Infinitives",
+        grammar_points=["gerunds-infinitives"],
+        vocabulary_set_ids=["media_society_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Correctly uses gerunds after enjoy, avoid, deny, suggest, consider",
+            "Correctly uses infinitives after want, decide, manage, refuse, hope",
+            'Identifies the meaning difference in "remember doing" vs "remember to do"',
+            "Uses a gerund after prepositions (interested in doing, used to doing)",
+            'Uses bare infinitive after "make" and "let"',
+        ],
+        default_weeks=1,
+        prerequisite_unit="b2-unit-4",
+    ),
+    CurriculumUnit(
+        id="b2-unit-6",
+        level="B2",
+        unit_number=6,
+        title="Reported Speech & Modal Perfects",
+        grammar_points=["reported-speech", "modal-perfects"],
+        vocabulary_set_ids=["news_events_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Applies full tense backshift (past perfect, would, could) in reported speech",
+            "Reports questions with if/whether and normal (non-inverted) word order",
+            "Uses must have, can't have, should have, could have correctly",
+            'Distinguishes "must have" (deduction) from "should have" (regret/criticism)',
+            "Reports a news story using appropriate verbs (said, told, added, denied)",
+        ],
+        default_weeks=2,
+        prerequisite_unit="b2-unit-5",
+    ),
+    CurriculumUnit(
+        id="b2-unit-7",
+        level="B2",
+        unit_number=7,
+        title="Concession, Contrast & Academic Discourse",
+        grammar_points=["concession-contrast-b2"],
+        vocabulary_set_ids=["academic_vocabulary_b2"],
+        lesson_types=["grammar", "vocabulary", "reading", "writing", "review"],
+        competency_checklist=[
+            "Uses although/even though to link two contrasting clauses in one sentence",
+            "Uses despite/in spite of with a noun or gerund (not a full clause)",
+            "Uses however/nevertheless to start a new contrasting sentence with a comma",
+            "Uses whereas to contrast two equal and opposite facts",
+            "Writes a structured argument with concession phrases to acknowledge counter-views",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b2-unit-6",
+    ),
+    CurriculumUnit(
+        id="b2-unit-8",
+        level="B2",
+        unit_number=8,
+        title="B2 Consolidation",
+        grammar_points=[
+            "past-perfect",
+            "second-conditional",
+            "third-conditional",
+            "gerunds-infinitives",
+            "reported-speech",
+            "modal-perfects",
+        ],
+        vocabulary_set_ids=["academic_vocabulary_b2", "workplace_b2", "media_society_b2"],
+        lesson_types=["reading", "writing", "review"],
+        competency_checklist=[
+            "Writes a 200-word discursive essay with a clear thesis, concessions, and conclusion",
+            "Demonstrates B2 grammar range: conditionals, passives, modal perfects, gerunds",
+            "Reports a short news story correctly using reported speech and modal perfects",
+            "Reads a 250-word text and correctly answers inferential comprehension questions",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b2-unit-7",
+    ),
+]
