+++ backend/app/data/en/curriculum_c1.py
@@ -1,0 +1,160 @@
+"""C1 curriculum units."""
+
+from __future__ import annotations
+
+from app.data.en._types import CurriculumUnit
+
+C1_UNITS: list[CurriculumUnit] = [
+    CurriculumUnit(
+        id="c1-unit-1",
+        level="C1",
+        unit_number=1,
+        title="Mixed Conditionals & Speculation",
+        grammar_points=["mixed-conditionals"],
+        vocabulary_set_ids=["abstract_concepts_c1"],
+        lesson_types=["grammar", "reading", "writing", "review"],
+        competency_checklist=[
+            "Construct mixed conditional sentences combining past and present time frames.",
+            "Speculate about past events and their present consequences.",
+            "Use abstract vocabulary to discuss hypothetical scenarios.",
+            "Evaluate the appropriateness of conditional structures in formal writing.",
+            "Identify mixed conditionals in authentic academic and literary texts.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="b2-unit-8",
+    ),
+    CurriculumUnit(
+        id="c1-unit-2",
+        level="C1",
+        unit_number=2,
+        title="Participle Clauses",
+        grammar_points=["participle-clauses"],
+        vocabulary_set_ids=["advanced_verbs_c1"],
+        lesson_types=["grammar", "reading", "writing"],
+        competency_checklist=[
+            "Use present and past participle clauses to reduce relative clauses.",
+            "Replace adverbial clauses with participial equivalents for conciseness.",
+            "Apply perfect participle clauses to express sequence of events.",
+            "Produce formal written texts using participle constructions naturally.",
+            "Distinguish between correct and ambiguous participle clause usage.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-1",
+    ),
+    CurriculumUnit(
+        id="c1-unit-3",
+        level="C1",
+        unit_number=3,
+        title="Hedging & Formal Register",
+        grammar_points=["hedging-language"],
+        vocabulary_set_ids=["formal_writing_c1"],
+        lesson_types=["grammar", "writing", "reading"],
+        competency_checklist=[
+            "Hedge claims appropriately in academic and professional writing.",
+            "Use formal connectors and discourse phrases accurately.",
+            "Distinguish between formal and informal register in written texts.",
+            "Produce a paragraph of formal writing using hedging language throughout.",
+            "Identify hedging devices in academic articles and reports.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-2",
+    ),
+    CurriculumUnit(
+        id="c1-unit-4",
+        level="C1",
+        unit_number=4,
+        title="Emphasis: Inversion & Cleft Sentences",
+        grammar_points=["inversion", "cleft-sentences"],
+        vocabulary_set_ids=["idioms_c1"],
+        lesson_types=["grammar", "writing", "review"],
+        competency_checklist=[
+            "Use subject–auxiliary inversion after negative and restrictive adverbials.",
+            "Construct it-cleft and wh-cleft sentences for emphasis.",
+            "Employ advanced idiomatic expressions naturally in spoken interaction.",
+            "Recognise and produce emphasis structures in formal speeches and essays.",
+            "Correct common errors in inversion and cleft constructions.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-3",
+    ),
+    CurriculumUnit(
+        id="c1-unit-5",
+        level="C1",
+        unit_number=5,
+        title="Ellipsis, Substitution & Textual Cohesion",
+        grammar_points=["ellipsis-substitution"],
+        vocabulary_set_ids=["academic_discourse_c1"],
+        lesson_types=["grammar", "reading", "writing"],
+        competency_checklist=[
+            "Apply ellipsis and substitution to avoid repetition in formal writing.",
+            "Use academic cohesion vocabulary to link arguments across paragraphs.",
+            "Analyse how cohesive devices contribute to text unity.",
+            "Produce an academic paragraph demonstrating effective use of ellipsis and reference.",
+            "Identify ellipsis and substitution in edited academic prose.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-4",
+    ),
+    CurriculumUnit(
+        id="c1-unit-6",
+        level="C1",
+        unit_number=6,
+        title="Advanced Relative Clauses & Critical Thinking",
+        grammar_points=["advanced-relative-clauses"],
+        vocabulary_set_ids=["critical_thinking_c1"],
+        lesson_types=["grammar", "reading", "writing", "review"],
+        competency_checklist=[
+            "Use non-defining, sentential, and reduced relative clauses correctly.",
+            "Express viewpoints using critical thinking vocabulary with precision.",
+            "Evaluate the validity and bias of arguments encountered in texts.",
+            "Integrate relative clauses fluidly into complex written sentences.",
+            "Debate a topic using counterargument and concession language.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-5",
+    ),
+    CurriculumUnit(
+        id="c1-unit-7",
+        level="C1",
+        unit_number=7,
+        title="Argumentation & Rhetoric",
+        grammar_points=["hedging-language", "inversion"],
+        vocabulary_set_ids=["debate_rhetoric_c1"],
+        lesson_types=["writing", "review", "grammar"],
+        competency_checklist=[
+            "Build a persuasive argument using assertion, evidence, and refutation.",
+            "Employ rhetorical devices (inversion, concession, emphasis) in speeches.",
+            "Use debate and rhetoric vocabulary to express agreement, challenge, and concession.",
+            "Write a 300-word opinion essay at C1 standard.",
+            "Deliver a 2-minute spoken argument on a complex topic.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-6",
+    ),
+    CurriculumUnit(
+        id="c1-unit-8",
+        level="C1",
+        unit_number=8,
+        title="C1 Consolidation",
+        grammar_points=[
+            "mixed-conditionals",
+            "participle-clauses",
+            "inversion",
+            "cleft-sentences",
+            "ellipsis-substitution",
+            "advanced-relative-clauses",
+            "hedging-language",
+        ],
+        vocabulary_set_ids=["abstract_concepts_c1", "academic_discourse_c1", "debate_rhetoric_c1"],
+        lesson_types=["grammar", "reading", "writing", "review"],
+        competency_checklist=[
+            "Produce a sustained piece of formal writing integrating all C1 grammar structures.",
+            "Demonstrate flexible control of register in both formal and semi-formal contexts.",
+            "Discuss abstract and complex topics fluently with appropriate hedging.",
+            "Identify and correct errors across all C1 grammar areas.",
+            "Pass a C1-level CEFR assessment covering all four skills.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-7",
+    ),
+]
