+++ backend/app/data/en/curriculum_c2.py
@@ -1,0 +1,129 @@
+"""C2 curriculum units."""
+
+from __future__ import annotations
+
+from app.data.en._types import CurriculumUnit
+
+C2_UNITS: list[CurriculumUnit] = [
+    CurriculumUnit(
+        id="c2-unit-1",
+        level="C2",
+        unit_number=1,
+        title="Discourse Markers & Text Cohesion",
+        grammar_points=["discourse-markers"],
+        vocabulary_set_ids=["nuanced_adjectives_c2"],
+        lesson_types=["grammar", "reading", "writing"],
+        competency_checklist=[
+            "Use a wide range of discourse markers to organise arguments at paragraph and essay level.",
+            "Identify how discourse markers signal logical relationships in sophisticated texts.",
+            "Employ nuanced C2 adjectives to add precision and depth to writing.",
+            "Produce a cohesive analytical essay of 400+ words using advanced connectors.",
+            "Evaluate the rhetorical effect of discourse marker choices in published writing.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c1-unit-8",
+    ),
+    CurriculumUnit(
+        id="c2-unit-2",
+        level="C2",
+        unit_number=2,
+        title="Nominalisation & Academic Style",
+        grammar_points=["nominalisation"],
+        vocabulary_set_ids=["formal_register_c2"],
+        lesson_types=["grammar", "writing", "reading"],
+        competency_checklist=[
+            "Convert verb and adjective phrases into noun phrases using nominalisation.",
+            "Produce academic writing with the impersonal, dense style characteristic of C2.",
+            "Use formal register vocabulary accurately in professional correspondence.",
+            "Identify and correct inappropriate register in academic texts.",
+            "Write an abstract or executive summary using nominalisation throughout.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c2-unit-1",
+    ),
+    CurriculumUnit(
+        id="c2-unit-3",
+        level="C2",
+        unit_number=3,
+        title="Idiomatic & Figurative Language",
+        grammar_points=["fronting-emphasis"],
+        vocabulary_set_ids=["idiomatic_expressions_c2"],
+        lesson_types=["vocabulary", "review", "reading"],
+        competency_checklist=[
+            "Use a wide range of idiomatic expressions naturally in spoken and written English.",
+            "Interpret figurative language in authentic literary and journalistic texts.",
+            "Employ fronting for rhetorical emphasis in spoken and written contexts.",
+            "Explain the meaning and cultural context of complex idioms.",
+            "Avoid inappropriate use of idioms in formal written registers.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c2-unit-2",
+    ),
+    CurriculumUnit(
+        id="c2-unit-4",
+        level="C2",
+        unit_number=4,
+        title="Fronting, Emphasis & Stylistic Devices",
+        grammar_points=["fronting-emphasis", "register-and-style"],
+        vocabulary_set_ids=["literary_devices_c2"],
+        lesson_types=["grammar", "writing", "reading"],
+        competency_checklist=[
+            "Use object fronting, adverbial fronting, and concessive fronting for effect.",
+            "Identify and deploy literary devices (metaphor, irony, juxtaposition) in writing.",
+            "Adapt writing style consciously for different purposes and audiences.",
+            "Analyse how stylistic choices shape meaning and tone in literary texts.",
+            "Produce a piece of creative or journalistic writing demonstrating stylistic variation.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c2-unit-3",
+    ),
+    CurriculumUnit(
+        id="c2-unit-5",
+        level="C2",
+        unit_number=5,
+        title="Critical Reading & Academic Writing",
+        grammar_points=["register-and-style"],
+        vocabulary_set_ids=["critical_analysis_c2"],
+        lesson_types=["reading", "writing", "review"],
+        competency_checklist=[
+            "Read and critically evaluate complex academic and professional texts.",
+            "Use critical analysis vocabulary (paradigm, empirical, discourse, posit) accurately.",
+            "Construct a structured critical analysis essay of 500+ words.",
+            "Interrogate the assumptions, bias, and methodology of academic arguments.",
+            "Demonstrate near-native precision of vocabulary and grammar in academic prose.",
+        ],
+        default_weeks=1,
+        prerequisite_unit="c2-unit-4",
+    ),
+    CurriculumUnit(
+        id="c2-unit-6",
+        level="C2",
+        unit_number=6,
+        title="C2 Consolidation",
+        grammar_points=[
+            "discourse-markers",
+            "nominalisation",
+            "fronting-emphasis",
+            "register-and-style",
+            "inversion",
+            "cleft-sentences",
+        ],
+        vocabulary_set_ids=[
+            "nuanced_adjectives_c2",
+            "formal_register_c2",
+            "idiomatic_expressions_c2",
+            "literary_devices_c2",
+            "critical_analysis_c2",
+        ],
+        lesson_types=["grammar", "reading", "writing", "review"],
+        competency_checklist=[
+            "Demonstrate mastery of all C2 grammar structures in extended writing tasks.",
+            "Use the full range of C2 vocabulary with precision and appropriate register.",
+            "Produce and deliver a coherent, sophisticated extended argument (spoken and written).",
+            "Critically evaluate a complex text and respond with a counter-argument in writing.",
+            "Achieve a score consistent with C2 proficiency on a Cambridge Proficiency-style assessment.",
+        ],
+        default_weeks=2,
+        prerequisite_unit="c2-unit-5",
+    ),
+]
