+++ backend/app/routers/reading.py
@@ -3,14 +3,13 @@ from __future__ import annotations
 import asyncio
 import logging
 
-from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Request
+from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Request, status
 from redis.asyncio import Redis
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.config import settings
-from app.core.database import AsyncSessionLocal, get_db
-from app.core.deps import get_current_user, require_subscription
+from app.core.database import get_db
+from app.core.deps import get_current_user, get_redis, require_subscription
 from app.core.limiter import limiter
 from app.models.study_plan import StudyPlan
 from app.models.user import User
@@ -31,22 +30,11 @@ from app.services.reading_service import (
     get_user_history,
     submit_attempt,
 )
+from app.utils.db import db_session
+from app.utils.redis import redis_client as _redis_client
 
 router = APIRouter(prefix="/api/reading", tags=["reading"])
 logger = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Local Redis dependency
-# ---------------------------------------------------------------------------
-
-
-async def get_redis():  # noqa: ANN201
-    redis = Redis.from_url(settings.REDIS_URL, decode_responses=True)
-    try:
-        yield redis
-    finally:
-        await redis.aclose()
 
 
 # ---------------------------------------------------------------------------
@@ -84,7 +72,7 @@ async def _get_user_level(user_id: int, db: AsyncSession) -> tuple[str, str]:
     )
     plan = result.scalar_one_or_none()
     if plan is None:
-        raise HTTPException(status_code=404, detail="no_study_plan")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="no_study_plan")
     return plan.cefr_level, plan.target_language
 
 
@@ -103,15 +91,14 @@ async def _background_generate(
     Creates its own DB session and Redis client (request resources are already closed).
     Releases the Redis lock in all cases (success or failure).
     """
-    redis_client = Redis.from_url(settings.REDIS_URL, decode_responses=True)
-    try:
-        async with AsyncSessionLocal() as db:
-            await generate_and_save_exercise(level, target_language, db)
-    except Exception:
-        logger.exception("reading: generation failed level=%s lang=%s", level, target_language)
-    finally:
-        await redis_client.delete(lock_key)
-        await redis_client.aclose()
+    async with _redis_client() as redis_conn:
+        try:
+            async with db_session() as db:
+                await generate_and_save_exercise(level, target_language, db)
+        except Exception:
+            logger.exception("reading: generation failed level=%s lang=%s", level, target_language)
+        finally:
+            await redis_conn.delete(lock_key)
 
 
 # ---------------------------------------------------------------------------
@@ -160,7 +147,9 @@ async def get_next_exercise(
     return ReadingNextResponse(available=False)
 
 
-@router.post("/generate", response_model=ReadingGeneratingResponse, status_code=202)
+@router.post(
+    "/generate", response_model=ReadingGeneratingResponse, status_code=status.HTTP_202_ACCEPTED
+)
 @limiter.limit("5/minute")
 async def generate_exercise(
     request: Request,
@@ -208,10 +197,14 @@ async def submit_reading_attempt(
     except ValueError as exc:
         detail = str(exc)
         if detail == "exercise_not_found":
-            raise HTTPException(status_code=404, detail="exercise_not_found") from exc
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="exercise_not_found"
+            ) from exc
         if detail == "already_attempted":
-            raise HTTPException(status_code=409, detail="already_attempted") from exc
-        raise HTTPException(status_code=400, detail=detail) from exc
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT, detail="already_attempted"
+            ) from exc
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=detail) from exc
 
     correct_answers = [
         CorrectAnswerOut(index=q["index"], correct=q["correct"]) for q in exercise.questions
