+++ backend/app/routers/stt.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, File, HTTPException, Request, UploadFile
+from fastapi import APIRouter, Depends, File, HTTPException, Request, UploadFile, status
 
 from app.core.deps import get_current_user
 from app.core.limiter import limiter
@@ -18,7 +18,9 @@ async def speech_to_text(
     """Proxy STT request to Whisper service. Returns transcribed text."""
     stt_service = getattr(request.app.state, "stt_service", None)
     if stt_service is None:
-        raise HTTPException(status_code=503, detail="STT service is not enabled")
+        raise HTTPException(
+            status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail="STT service is not enabled"
+        )
     audio_bytes = await audio.read()
     if len(audio_bytes) > 50 * 1024 * 1024:  # 50 MB
         raise HTTPException(status_code=413, detail="Audio file too large (max 50 MB)")
